pyproject.toml
[project]
name = "knowledge-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "google-api-python-client>=2.189.0",
    "google-auth-oauthlib>=1.2.4",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-openai>=1.1.8",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "slack-bolt>=1.27.0",
]
